refactor: common files for globals
