+++ src/views/zesty/Product.js
@@ -42,8 +42,8 @@ import { AlgoSearch } from 'views/Docs/AlgoSearch';
 import { useZestyStore } from 'store';
 import GetDemoSection from 'revamp/ui/GetDemoSection';
 import { ZestyMarkdownParser } from 'components/markdown-styling/ZestyMarkdownParser';
-import { TreeNavigation } from 'components/Global/TreeNavigation';
-import { TableOfContent } from 'components/Global/TableOfContent';
+import { TreeNavigation } from 'components/globals/TreeNavigation';
+import { TableOfContent } from 'components/globals/TableOfContent';
 
 // main file
 const Product = (props) => {
